recipes

+++ controllers/recipeController.js
@@ -1,12 +1,10 @@
 const mongoose = require('mongoose');
 
-const catchErrors = (asyncFunction) => (...args) => asyncFunction(...args).catch(console.error);
-
 const { RecipeSchema, recschema } = require('../models/recipeModel');
 
 // create and save new recipe
 exports.create = async (req, res) => {
-	// console.log(request.body);
+	// console.log(req.body);
 	const recipe = new RecipeSchema({
 		title: req.body.title,
 		categories: req.body.categories,
@@ -14,54 +12,52 @@ exports.create = async (req, res) => {
 		description: req.body.description,
 		author: req.body.author
 	});
-	await recipe.save();
-	res.redirect('/');
-};
+	const verif = recschema.validate(recipe);
 
-{
-	/*exports.create = catchErrors(upload.single('image'), async (req, res) => {
-	const body = req.body;
-	const verif = schema.validate(body);
-	// validate request
 	if (verif.error) {
 		res.status(400).send(verif.error.details[0].message);
 		return;
 	}
-	// new user
-	const recipe = new RecipeSchema(body);
-	// save recipe
-	await recipe.save(recipe);
+	await recipe.save();
 	res.redirect('/');
-});*/
-}
+};
 
 // retrive and return all recipes or single
-exports.find = catchErrors(async (req, res) => {
+exports.find = async (req, res) => {
 	if (req.query.id) {
 		const id = req.query.id;
 
-		const data = await RecipeSchema.findById(id);
-
-		if (!data) {
-			res.status(404).send({ message: 'Not found recipe with id ' + id });
-		} else {
-			res.send(data);
-		}
+		await RecipeSchema.findById(id)
+			.then((data) => {
+				if (!data) {
+					res.status(404).send('Not found recipe with id ' + id);
+				} else {
+					res.send(data);
+				}
+			})
+			.catch((err) => {
+				res.status(500).send('Erro retrieving user with id ' + id);
+			});
 	} else {
-		const recipe = await RecipeSchema.find();
-		res.send(recipe);
+		await RecipeSchema.find()
+			.then((recipe) => {
+				res.send(recipe);
+			})
+			.catch((err) => {
+				res.status(500).send('Error Occurred while retriving user information');
+			});
 	}
-});
+};
 
 exports.update = async (req, res) => {
-	const id = req.params.id;
+	const id = req.query.id;
 
 	const verifID = mongoose.Types.ObjectId.isValid(id);
 
 	let body = req.body;
 	let file = req.file;
 
-	const verif = recschema.validate(body);
+	const verif = recschema.validate(body, file);
 	if (!verifID) {
 		res.status(400).send('id non conforme !');
 		return;
@@ -72,41 +68,48 @@ exports.update = async (req, res) => {
 		return;
 	}
 
-	let upData = await RecipeSchema.findById(id);
+	await RecipeSchema.findById(id)
+		.then((upData) => {
+			if (!upData) {
+				res.status(404).send("aucun enregistrement trouvé pour l'id " + id);
+				return;
+			}
 
-	if (!upData) {
-		res.status(404).send("aucun enregistrement trouvé pour l'id " + id);
-		return;
-	}
+			upData.image = file.filename;
+			upData.title = body.title;
+			upData.categories = body.categories;
+			upData.description = body.description;
+			upData.author = body.author;
+			upData.timeCreated = body.timeCreated;
 
-	upData.image = file.filename;
-	upData.title = body.title;
-	upData.categories = body.categories;
-	upData.description = body.description;
-	upData.author = body.author;
-	upData.timeCreated = body.timeCreated;
-
-	await upData.save();
-	res.redirect('/');
+			upData.save();
+			res.redirect('/');
+		})
+		.catch((err) => {
+			res.status(500).send('Error Update recipe information');
+		});
 };
 
 // delete a recipe with specified recipe id
-exports.delete = (req, res) => {
+exports.delete = async (req, res) => {
 	const id = req.params.id;
 
-	RecipeSchema.findByIdAndDelete(id)
+	const verifID = mongoose.Types.ObjectId.isValid(id);
+
+	if (!verifID) {
+		res.status(400).send("l'id transmis n'est pas conforme");
+		return;
+	}
+
+	await RecipeSchema.findByIdAndDelete(id)
 		.then((data) => {
 			if (!data) {
-				res.status(404).send({ message: `Cannot Delete with id ${id}. Maybe id is wrong` });
+				res.status(404).send(`Cannot Delete with id ${id}. Maybe id is wrong`);
 			} else {
-				res.send({
-					message: 'Recipe was deleted successfully!'
-				});
+				res.redirect('/view-recipes');
 			}
 		})
 		.catch((err) => {
-			res.status(500).send({
-				message: 'Could not delete Recipe with id=' + id
-			});
+			res.status(500).send('Could not delete recipe with id=' + id);
 		});
 };
